+++ src/main/java/com/booleanuk/core/Filling.java
@@ -2,7 +2,7 @@ package com.booleanuk.core;
 
 public class Filling extends Item {
 
-    public Filling(String sku, Double price, String name, String variant){
+    public Filling(String sku, Integer price, String name, String variant){
         super(sku, price, name, variant);
     }
 }
